+++ bin/check-azurerm-service-bus-topic-size.rb
@@ -127,8 +127,7 @@ class CheckAzureRMServiceBusTopicSize < Sensu::Plugin::Check::CLI
     else
       ok message
     end
-
-  rescue => e
+  rescue StandardError => e
     puts "Error: exception: #{e}"
     critical
   end