+++ bin/check-azurerm-virtual-machines-usage.rb
@@ -113,8 +113,7 @@ class CheckAzureRMVMUsage < Sensu::Plugin::Check::CLI
     else
       ok message
     end
-
-  rescue => e
+  rescue StandardError => e
     puts "Error: exception: #{e}"
     critical
   end