+++ bin/check-azurerm-virtual-network-gateway-connected.rb
@@ -99,13 +99,12 @@ class CheckAzureRMVirtualNetworkGatewayConnected < Sensu::Plugin::Check::CLI
     inbound = result.ingress_bytes_transferred
     outbound = result.egress_bytes_transferred
     message = "State is '#{connection_state}'. Usage is #{inbound} in / #{outbound} out"
-    if result.connection_status.casecmp('connected') == 0
+    if result.connection_status.casecmp('connected').zero?
       ok message
     else
       critical message
     end
-
-  rescue => e
+  rescue StandardError => e
     puts "Error: exception: #{e}"
     critical
   end